ship a forward compatible annotation route loader
In previous versions of EasyAdminBundle, the loading of annotated routes
was handled by the SensioFrameworkExtraBundle. With version 5.2 of the
SensioFrameworkExtraBundle their route annotation loader was deprecated
in favour of the implementation shipped with the Symfony FrameworkBundle
3.4+.

This commit provides a BC layer by wiring up the implementation of the
annotated route controller for older versions of the FrameworkBundle in
the same manner the Symfony core uses nowadays.